+++ src/ActionGameCore/Assets/TPSCameraControls/Scripts/Runtime/TPSCameraController.cs
@@ -1,4 +1,5 @@
 #if SUPPORT_INPUTSYSTEM
+using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
@@ -13,7 +14,10 @@ namespace FPSCameraControls
         }
 
         [SerializeField]
-        private InputActionReference _lookActionReference = null;
+        private InputActionReference _deltaActionReference = null;
+
+        [SerializeField]
+        private InputActionReference _continuousActionreference = null;
 
         [SerializeField]
         public Transform Target = null;
@@ -28,7 +32,7 @@ namespace FPSCameraControls
         public float Distance = 5.0f;
 
         [SerializeField]
-        public float Sensitivity = 0.1f;
+        public float Sensitivity = 1.0f;
 
         [SerializeField]
         public float AngleMin = -90.0f;
@@ -43,13 +47,19 @@ namespace FPSCameraControls
         public LayerMask WallLayerMask = ~0;
 
         [SerializeField]
+        public int SmoothingFrameCount = 2;
+
+        [SerializeField]
         private bool _lockAndHideCursor = true;
 
         private Vector3 _lookAngles;
+        private List<Vector2> _deltaPositions = new List<Vector2>();
+        private List<float> _deltaTimes = new List<float>();
 
         private void Start()
         {
-            _lookActionReference?.action.Enable();
+            _deltaActionReference?.action.Enable();
+            _continuousActionreference?.action.Enable();
             if (_lockAndHideCursor)
             {
                 Cursor.visible = false;
@@ -64,14 +74,42 @@ namespace FPSCameraControls
                 return;
             }
 
-            var rotation = Target.rotation;
-            if (_lookActionReference != null)
+            var deltaAngle = Vector2.zero;
+            if (_deltaActionReference != null)
             {
-                var look = _lookActionReference.action.ReadValue<Vector2>();
-                _lookAngles.x = Mathf.Clamp(_lookAngles.x - look.y * Sensitivity, -AngleMax, -AngleMin); // Look up and down
-                _lookAngles.y = (_lookAngles.y + look.x * Sensitivity) % 360; // Look left and right
-                rotation = Quaternion.Euler(_lookAngles);
+                _deltaPositions.Add(_deltaActionReference.action.ReadValue<Vector2>());
+                _deltaTimes.Add(Time.deltaTime);
+                SmoothingFrameCount = Mathf.Max(1, SmoothingFrameCount);
+                if (_deltaPositions.Count > SmoothingFrameCount)
+                {
+                    _deltaPositions.RemoveRange(0, _deltaPositions.Count - SmoothingFrameCount);
+                    _deltaTimes.RemoveRange(0, _deltaTimes.Count - SmoothingFrameCount);
+                }
+                var totalPosition = Vector2.zero;
+                var totalTime = 0f;
+                for (int i = 0; i < _deltaPositions.Count; i++)
+                {
+                    totalPosition += _deltaPositions[i];
+                    totalTime += _deltaTimes[i];
+                }
+                var deltaPositionAverage = totalPosition / totalTime;
+                var smoothDeltaPosition = deltaPositionAverage * Time.deltaTime;
+
+                const float DpiAverage = 96;
+                var dpi = Screen.dpi == 0 ? DpiAverage : Screen.dpi;
+                const float InchForTurn = 13;
+                deltaAngle += smoothDeltaPosition / dpi / InchForTurn * 180;
             }
+
+            if (_continuousActionreference != null)
+            {
+                const float SecondsForTurn = 1.0f;
+                deltaAngle += _continuousActionreference.action.ReadValue<Vector2>() * Time.deltaTime / SecondsForTurn * 180;
+            }
+
+            _lookAngles.x = Mathf.Clamp(_lookAngles.x - deltaAngle.y * Sensitivity, -AngleMax, -AngleMin); // Look up and down
+            _lookAngles.y = (_lookAngles.y + deltaAngle.x * Sensitivity) % 360; // Look left and right
+            var rotation = Quaternion.Euler(_lookAngles);
 
             var offsetRotation = OffsetRotation == OffsetRotationType.FullRotation ? rotation : Quaternion.AngleAxis(_lookAngles.y, Vector3.up);
             var pivot = Target.position + offsetRotation * Offset;